requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
google-genai==1.73.1
mcp==1.27.0
python-dotenv==1.2.2
